+++ DecenaSoluciones.POS.API/Services/Contracts/IProductService.cs
@@ -12,7 +12,7 @@ namespace DecenaSoluciones.POS.API.Services
         Task<ProductViewModel> UpdateProduct(int id, AddEditProduct product);
         Task<bool> UpdateInventary(List<UpdateInventory> inventoryItems);
         Task<int> RemoveProduct(int id);
-        Task<ProductViewModel> UpdateProductStock(int id, int quantity);
+        Task<ProductViewModel> UpdateProductStock(int id, decimal quantity);
         Task<LastSaleXProductViewModel> GetLastSaleXProduct(int id);
     }
 }
