+++ DecenaSoluciones.POS.Shared/Dtos/DashboardViewModel.cs
@@ -18,7 +18,7 @@ namespace DecenaSoluciones.POS.Shared.Dtos
     public class SoldProductQuantityViewModel
     {
         public string ProductName { get; set; }
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
 
     }
 }
