+++ DecenaSoluciones.POS.Shared/Dtos/ProductViewModel.cs
@@ -11,7 +11,7 @@ namespace DecenaSoluciones.POS.Shared.Dtos
         public string Description { get; set; }
         public int? MaintenancePeriods { get; set; }
         public int? WarrantyTime { get; set; }
-        public int stock { get; set; }
+        public decimal stock { get; set; }
         public decimal Price { get; set; }
         public decimal Cost { get; set; }
         public bool Assignable { get; set; }
