+++ DecenaSoluciones.POS.Shared/Dtos/SoldProductsReportViewModel.cs
@@ -6,8 +6,8 @@ namespace DecenaSoluciones.POS.Shared.Dtos
     {
         public string Code { get; set; }
         public string Description { get; set; }
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public decimal SalesTotal { get; set; }
-        public int Stock { get; set; }
+        public decimal Stock { get; set; }
     }
 }
